# add a sublist primop