@balsoft 2023-05-04
> I'm also a bit curious about the exact semantics: Currently sublist start len l will throw if min < 0 but shrink things if start >= length(l). Shouldn't we make it also throw if start >= length(l) for symmetry?

Fair point. TBH I'd rather make it a total function.