@balsoft 2023-05-04
It's now total, and behaves symmetrically.