@balsoft 2023-05-04
Trivial benchmark:

```nix
rec {
  lib = (import <nixpkgs> {}).lib;
  list = builtins.fromJSON (builtins.readFile ./list.json);
  sublists = builtins.genList (n: lib.sublist n (builtins.elemAt list n) list) (builtins.length list);
  sublists-new = builtins.genList (n: builtins.sublist n (builtins.elemAt list n) list) (builtins.length list);
}
```

where `list.json` contains 3000 entries

Results:

```
# Old
./src/nix/nix --extra-experimental-features nix-command eval -f ~/bench/sublists.nix sublists  1.77s user 0.27s system 84% cpu 2.428 total
# New
./src/nix/nix --extra-experimental-features nix-command eval -f ~/bench/sublists.nix sublists-new  0.34s user 0.13s system 55% cpu 0.854 total
```

More realistic benchmark (kinda inefficient merge sort):

```nix
let
  sublist =
    # Index at which to start the sublist
    start:
    # Number of elements to take
    count:
    # Input list
    list:
    let len = builtins.length list; in
    builtins.genList
      (n: builtins.elemAt list (n + start))
      (if start >= len then 0
       else if start + count > len then len - start
       else count);

  take = count: sublist 0 count;
  drop = count: list: sublist count (builtins.length list) list;

  merge = xs: ys: let x = builtins.head xs; y = builtins.head ys; in if builtins.length xs == 0 then ys else if builtins.length ys == 0 then xs else if x <= y then [x] ++ merge (builtins.tail xs) ys else [y] ++ merge xs (builtins.tail ys);
  msort = xs : if builtins.length xs == 0 then [] else if builtins.length xs == 1 then xs else merge (msort (firstHalf xs)) (msort (secondHalf xs));
  firstHalf = xs: let n = builtins.length xs; in take (builtins.div n 2) xs;
  secondHalf = xs: let n = builtins.length xs; in drop (builtins.div n 2) xs;

  example = builtins.fromJSON (builtins.readFile ./list.json);
in
  msort example
```

As-is:

```
./src/nix/nix --extra-experimental-features nix-command eval -f  > /dev/null  8.01s user 0.90s system 100% cpu 8.911 total

```

Using `builtins.sublist` instead:

```
./src/nix/nix --extra-experimental-features nix-command eval -f  > /dev/null  7.53s user 0.73s system 100% cpu 8.260 total
```

All results seem to be fairly consistent run-to-run

In short,

| Benchmark | Old | New | % improvement |
|-|-|-|-|
| Simple | 2.428 | 0.854 | 64% |
| Merge-sort | 8.911 | 8.260 | 7% |