@thufschmitt 2023-05-04
I tried some quick-and-dirty benchmarks on nixpkgs running `nix eval --file . firefox.outPath --option eval-cache false` on a checkout of nixpkgs both pristine and with the implementation of `lib.sublist` replaced by `builtins.sublist`. Hyperfine shows me essentially no difference between both (I ran this a couple time, and the timing difference is largely within the error margin). @balsoft do you confirm that? Because if so it would mean that this isn't really worth it in practice (but that contradicts @cleverca22 's previous benchmarks, so I might have done something wrong)