@thufschmitt 2023-05-04
More precisely: I'm in a checkout of NixOS/nixpkgs@3cdb57d5ad771a9e9ac6eb258f89de201d928a11 with the patch below applied

<details><summary>nixpkgs.patch</summary>
<p>

```diff
diff --git a/lib/lists.nix b/lib/lists.nix
index 2186cd4a79f..e0bd403fc05 100644
--- a/lib/lists.nix
+++ b/lib/lists.nix
@@ -594,6 +594,9 @@ rec {
        => [ ]
   */
   sublist =
+    if builtins.getEnv "NIX_USE_BUILTIN_SUBLIST" == 1
+    then builtins.sublist
+    else
     # Index at which to start the sublist
     start:
     # Number of elements to take
```

</p>
</details> 

and I'm running

```
nix shell nixpkgs#hyperfine github:tweag/nix/add-sublist --command hyperfine --warmup 1 "NIX_USE_BUILTIN_SUBLIST={bltSublist} nix eval --file . firefox.outPath --option eval-cache false" --parameter-list bltSublist 0,1
```

Which gives me

```console
$ nix shell nixpkgs#hyperfine github:tweag/nix/add-sublist --command hyperfine --warmup 1 "NIX_USE_BUILTIN_SUBLIST={bltSublist} nix eval --file . firefox.outPath --option eval-cache false" --parameter-list bltSublist 0,1
Benchmark 1: NIX_USE_BUILTIN_SUBLIST=0 nix eval --file . firefox.outPath --option eval-cache false
  Time (mean ± σ):      2.955 s ±  0.045 s    [User: 1.982 s, System: 0.530 s]
  Range (min … max):    2.908 s …  3.036 s    10 runs
 
Benchmark 2: NIX_USE_BUILTIN_SUBLIST=1 nix eval --file . firefox.outPath --option eval-cache false
  Time (mean ± σ):      2.967 s ±  0.045 s    [User: 1.987 s, System: 0.536 s]
  Range (min … max):    2.915 s …  3.065 s    10 runs
 
Summary
  'NIX_USE_BUILTIN_SUBLIST=0 nix eval --file . firefox.outPath --option eval-cache false' ran
    1.00 ± 0.02 times faster than 'NIX_USE_BUILTIN_SUBLIST=1 nix eval --file . firefox.outPath --option eval-cache false'
```