@balsoft 2023-05-04
I get similar results, and I wouldn't expect much difference anyways since `sublist` isn't that often used to evaluate `firefox`. Perhaps @cleverca22 was doing something more interesting?