@thufschmitt 2023-05-04
Heck :/ I was hoping that this was useful because stdenv somehow ended-up calling `sublist` a lot. But if not it might not be worth it :/

I also tried by instantiating a nixos test instead of firefox to see whether there was a difference, but there's nothing meaningful either